rename and improve docs
