+++ module_text_llm/module_text_llm/config.py
@@ -2,13 +2,14 @@ from pydantic import BaseModel, Field
 
 from athena import config_schema_provider
 from module_text_llm.helpers.models import ModelConfigType, DefaultModelConfig
-from .prompts.suggest_feedback_basic import system_template, human_template
+from module_text_llm.prompts.generate_suggestions import system_template, human_template
 
 
-class BasicPrompt(BaseModel):
+class GenerateSuggestionsPrompt(BaseModel):
     """\
-Features available: **{problem_statement}**, **{example_solution}**, **{grading_instructions}**, **{submission}**, **{max_points}**, **{bonus_points}**
-**{problem_statement}** or **{example_solution}** might be omitted if the input is too long.\
+Features available: **{problem_statement}**, **{example_solution}**, **{grading_instructions}**, **{max_points}**, **{bonus_points}**, **{submission}**
+
+_Note: **{problem_statement}**, **{example_solution}**, or **{grading_instructions}** might be omitted if the input is too long._\
 """
     system_message: str = Field(default=system_template,
                                 description="Message for priming AI behavior and instructing it what to do.")
@@ -20,7 +21,7 @@ class BasicApproachConfig(BaseModel):
     """This approach uses a LLM with a single prompt to generate feedback in a single step."""
     max_input_tokens: int = Field(default=3000, description="Maximum number of tokens in the input prompt.")
     model: ModelConfigType = Field(default=DefaultModelConfig()) # type: ignore
-    prompt: BasicPrompt = Field(default=BasicPrompt())
+    prompt: GenerateSuggestionsPrompt = Field(default=GenerateSuggestionsPrompt())
 
 
 @config_schema_provider
